requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
bcrypt==4.0.1
PyMySQL==1.1.1
python-dotenv==1.1.0